feat: CAT-98 FeedImageService를 구현체와 인터페이스로 분리
